Chronicles of Pumma is a 2D Open Source Game developed using the awesome LibGDX Framework.
Chronicles of Pumma is a 2D Mario like Open Source android game
developed using the LibGDX Framework by Digital Wolf Studio. Pumma is an
adventurous guy who is in a wonderland and has to collect more and more
gems and eggs to maintain his health. He has to prevent himself from
poisonous snakes and dreadful dragons during his journey. He has to find
the key so as to unlock the door to the next Level. Just go slow and be
wise ...

Feel free to use the source code for learning or making a commercial
game.

The game is published on Google Play at

https://play.google.com/store/apps/details?id=com.digitalwolf.adventuresOfPumma